linear-core: f32 recurrent state — SIMD-able GDN, heads across the pool
The GDN condensate ran in f64: 4x the traffic of the vendor's own dtype
(mamba_ssm_dtype: float32 in the source configs), no SIMD, one thread,
two full passes over S per head, fresh kv/o vecs per head. Now: state
stored f32 (vmf_phase keeps f64 math per cell at half the storage), the
two S passes regrouped to read-twice/write-once with elementwise-in-dj
loops LLVM auto-vectorizes (fmla/FMA), heads fan out across the pool,
per-worker scratch from the attention freelists. GDN oracle stays green
at 1e-3 (the fixture itself is f32 HF math); Bonsai-27B greedy output
unchanged (84*3/2 -> 252 -> 126).

Honest end-to-end verdict: on the q1-quantized 27B the token time is
currently ~12 GB/s of q1 bit-unpack ALU (449 dispatches/token and the
GDN state both sit in its shadow), so this lands as correctness+
capacity (state RAM halved) until the q1 kernel gets cheaper per
weight. Also: q1 SDOT kernel reshaped to 4 tiles/iteration (vpaddq tree
+ batched scale FMA instead of per-tile vaddvq + scalar chain).

+++ crates/cortiq-engine/src/kv_cache.rs
@@ -97,9 +97,9 @@ pub struct LayerKvCache {
     pub num_kv_heads: usize,
     pub head_dim: usize,
     /// Linear-core condensate S (vmf_phase), f64; empty on full layers.
-    pub linear_state: Vec<f64>,
+    pub linear_state: Vec<f32>,
     /// Tentative lane-2 state during speculative verify.
-    pub linear_scratch: Vec<f64>,
+    pub linear_scratch: Vec<f32>,
     /// O(1) Nyström override (None = plain cache attention).
     pub o1: Option<O1State>,
 }
@@ -681,7 +681,7 @@ impl LayerKvCache {
             // O(1) recurrent state of linear-core layers (vmf_phase/GDN):
             // constant in context, but real memory — the honest "KV+state"
             // line must count it (a pure-linear model reported 0 before).
-            + self.linear_state.len() * std::mem::size_of::<f64>()
+            + self.linear_state.len() * std::mem::size_of::<f32>()
             // O(1) Nyström state (window + sinks + skeleton) — same
             // discipline: constant in context, but real memory.
             + self.o1_memory_bytes()
